Fixed p2p errors, fixed room broadcast bugs and text messaging works now
